requirements.txt
streamlit==1.36.0
langchain-groq==0.1.9
python-dotenv==1.0.1
reportlab==4.2.2
pypdf==4.2.0
Pillow==10.4.0
requests==2.32.3